+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #3050 - Garbage Collector - Removed unused package for compatible for AEMaaCS
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->
 
+### Added
+
+- #2929 - Reports - added a new Report column type for Predicted Tags (Assets)
+
 ## 5.6.0 - 2023-02-02
 
 ### Added